server/requirements.txt
APScheduler==3.6.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
elasticsearch==7.5.1
Flask==1.1.1
Flask-Cors==3.0.8
geographiclib==1.50
geojson==2.5.0
geopy==1.21.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
mailjet-rest==1.3.3
MarkupSafe==1.1.1
oauth2==1.9.0.post1
praw==6.5.1
prawcore==1.0.1
pyowm==2.10.0
pytz==2019.3
requests==2.21.0
six==1.14.0
tzlocal==2.0.0
update-checker==0.16
urllib3==1.24.2
websocket-client==0.57.0
Werkzeug==0.16.0
youtube-python==1.0.13
